pyproject.toml
[tool.poetry]
name = "Projeto"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
django = "5.0"
django-allauth = "65.3.0"
django-crispy-forms = "2.3"
django-geojson = "4.1.0"
pillow = "11.0.0"
sqlparse = "0.5.2"
tzdata = "2024.2"
crispy-bootstrap4 = "^2024.10"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
channels = "^4.2.1"
channels-redis = "^4.2.1"
redis = "^5.2.1"
numpy = "^2.2.4"
mssql-django = "^1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
